Shorthand for curried functions
Propagate capture sets to the right in curried functions. Example:

    {x} A -> B -> C

is a shorthand for

    {x} A -> {x} B -> C

or:

    (x: {*} A) -> B -> C

is a shorthand for

    (x: {*} A) -> {x} B -> C

or:

    ({*} A) -> B -> C

is a shorthand for

    (x$0: {*} A) -> {x$0} B -> C

Also: allow empty capture sets in types

This gives a more convenient override to disable capture set propagation
in curried types than wrapping in a type alias. E.g. compare

    {x} A -> {} B -> C

with

    {x} A -> Protect[B -> C]

where

    type Protect[X] = X

Also: refactoring to move setup code from Rechecker and CheckCaptures into a joint
class cc.Setup.
